Added column view control by security level
Now it is possible to indicate a level of security to each of the columns that is compared with the level of security of the user. If the user's level is not enough, the column is not shown.

IMPORTANT: Only administrator users can set or modify the security level of the columns.