Changed the structure of the rows type head
The "head" type rows now become "statistics" type, keeping their operation unchanged.
The RowItemFooter class is renamed by RowItemCards more in line with its new function, leaving it ready for the next update where it will be possible to indicate if the information panels will be shown in the header or in the footer.