Added possibility to disable tabs in Panel Controller
The new protected "checkActiveView" method that is executed during the data loading of the views and indicates whether the view will be available to the user. By default, the status depends on whether the main view of the controller has data, that is, if a new record is not being registered.

Also, the "icons" property has been restructured. This property has been removed and the "settings" property has been added which contains, for each view, the list of configuration properties such as icon and active.

A "getSettings" method has been added, which returns the configuration value for the indicated view and property.