Add enterMoves to the event handler and other improvements
Now the enterMoves keyboard event is controlled from the event manager.
The isAutoComplete parameter is added to the event beforeChanges so that the child classes know if the modification is due to the selection of a value from the auto-completed list or by user input.
The column delimiter of the autocomplete fields is changed by double separator to avoid cases in which the user enters a single separator.