Change getByDate by loadFromDate
Modified the getByDate method by loadFromDate. Now the method loads the exercise data on the existing model for the indicated date. It is necessary that the exercise model has the company identifier on which the search is carried out.

Corrected the calls to old method in bussines documents and accounting entry.

Changed the "abierto" method to "isOpened"